test(chart_state): add test for legendItemsTooltipValues
